requirements.txt
scikit-learn
tensorflow==2.8.2
matplotlib
keras
pandas-gbq==0.17.0
google-cloud-storage
google-cloud-bigquery
